+++ lib/features/booking/logic/booking_filter_state.dart
@@ -1,0 +1,7 @@
+part of 'booking_filter_cubit.dart';
+
+@freezed
+class BookingFilterState with _$BookingFilterState {
+  const factory BookingFilterState.updated(Map<BookingFilters, bool> filters) =
+      _Updated;
+}